requirements.txt
grpcio
grpcio-tools
protobuf
streamlit